agridat {agridat}	R Documentation

Datasets from agricultural experiments

Description

This package contains datasets from publications relating to agriculture, including field crops, tree crops, animal studies, and a few others.

Details

If you use these data, please cite both the agridat package and the original source of the data.

Abbreviations in the 'other' column include: xy = coordinates, pls = partial least squares, rsm = response surface methodology, row-col = row-column design, ts = time series,

Uniformity trials with a single genotype

name 	 dimensions 	 other 	 model

ansari.wheat.uniformity 	 96 x 8 	 xy

bailey.cotton.uniformity 	 8 x 20 	 xy, 5 envts

baker.barley.uniformity 	 3 x 19 	 xy, 10 years

baker.strawberry.uniformity 	 10 x 50 	 xy

baker.wheat.uniformity 	 12 x 12 	 xy

bancroft.peanut.uniformity 	 6 x 18 	 xy, 2 blocks

batchelor.apple.uniformity 	 8 x 28 	 xy

batchelor.lemon.uniformity 	 14 x 16 	 xy

batchelor.navel1.uniformity 	 20 x 50 	 xy

batchelor.navel2.uniformity 	 15 x 33 	 xy

batchelor.valencia.uniformity 	 12 x 20 	 xy

batchelor.walnut.uniformity 	 10 x 28 	 xy

beckett.maize.uniformity 	 6 x 19 	 xy

bose.multi.uniformity 	 15 x 26 	 xy, 3 years

bradley.multi.uniformity 	 10 x 11 	 xy

bryan.corn.uniformity 	 6 x 48 	 xy

christidis.cotton.uniformity 	 16 x 16 	 xy, 4 blocks

christidis.wheat.uniformity 	 12 x 24 	 xy

coombs.rice.uniformity 	 3 x 18 	 xy

dasilva.soybean.uniformity 	 24 x 48 	 xy

davies.pasture.uniformity 	 19 x 40 	 xy

day.wheat.uniformity 	 31 x 100 	 xy

draper.safflower.uniformity 	 18 x 16 	 xy, 2 expt 	 smith

ducker.groundnut.uniformity 	 44 x 5 	 xy

eden.tea.uniformity 	 12 x 12 	 xy

evans.sugarcane.uniformity 	 142 x 5 	 xy

forster.wheat.uniformity 	 10 x 16 	 xy

garber.multi.uniformity 	 6 x 45 	 xy, 5 years/crops

gomez.rice.uniformity 	 18 x 36 	 xy 	 aov,smith

gorski.oats.uniformity 	 10 x 20, 15 x 20 	 xy

goulden.barley.uniformity 	 48 x 48 	 xy

hansen.multi.uniformity 	 6 x 5, 16 x 8 	 xy, 4-5 yr 	 splom

haritonenko.sugarbeet.uniformity 	 4 x 104 	 xy

harris.multi.uniformity 	 2 x 23 	 xy, 23 crops 	 corrgram

hartman.tomato.uniformity 	 12 x 32 	 xy

heath.cabbage.uniformity 	 6 x 8 	 xy

heath.radish.uniformity 	 10 x 10 	 xy

holtsmark.timothy.uniformity 	 6 x 40 	 xy

hutchinson.cotton.uniformity 	 36 x 42 	 xy

igue.sugarcane.uniformity 	 36 x 42 	 xy

immer.sugarbeet.uniformity 	 10 x 60 	 xy, 3 traits, 2 yr

iyer.wheat.uniformity 	 25 x 80 	 xy

jegorow.oats.uniformity 	 8 x 30 	 xy

jones.corn.uniformity 	 12 x 12 	 xy

jurowski.wheat.uniformity 	 4 x 40 	 xy

kadam.millet.uniformity 	 8 x 20 	 xy, 2 expts

kalamkar.potato.uniformity 	 6 x 96 	 xy

kalamkar.wheat.uniformity 	 16 x 80 	 xy, 2 traits

kempton.barley.uniformity 	 7 x 28 	 xy

kerr.sugarcane.uniformity 	 12 x 12 	 xy 	, 4 fields

khan.brassica.uniformity 	 18 x 18 	 xy, 2 fields

khin.rice.uniformity 	 30 x 36 	 xy

kirk.potato.uniformity 	 6 x 6 	 xy

kiesselbach.oats.uniformity 	 3 x 69 	 xy

kotowski.potato.uniformity 	 12 x 4, 26 x 4 	 xy

kristensen.barley.uniformity 	 22 x 11 	 xy

kulkarni.sorghum.uniformity 	 4 x 40 	 xy, 3 years

lander.multi.uniformity 	 5 x 39 	 xy, 4 years

larsen.timothy.uniformity 	 12 x 80 	 xy

laycock.tea.uniformity 	 6 x 9 	 xy, 2 locations

lehmann.millet.uniformity 	 6 x 22 	 xy, 3 years

lehmann.rice.uniformity 	 2 x 17 	 xy, 4 years

lessman.sorghum.uniformity 	 46 x 60 	 xy

li.millet.uniformity 	 6 x 100 	 xy

loesell.bean.uniformity 	 21 x 90 	 xy

lord.rice.uniformity 	 5 x 14 	 xy, 8 fields

love.cotton.uniformity 	 16 x 10 	 xy

love.sugarcane.uniformity 	 20 x 20 	 xy

lyon.potato.uniformity 	 34 x 6 	 xy

magistad.pineapple.uniformity 	 5 x 5 	 xy

masood.rice.uniformity 	 12 x 24 	 xy

mcclelland.corn.uniformity 	 2 x 44 	 xy

mckinstry.cotton.uniformity 	 24 x 20 	 xy

mercer.mangold.uniformity 	 10 x 20 	 xy

mercer.wheat.uniformity 	 25 x 20 	 xy, 2 traits 	 spplot

montgomery.wheat.uniformity 	 14 x 16 	 xy, 2 years 	 lm

moore.polebean.uniformity 	 12 x 12 	 xy

moore.bushbean.uniformity 	 24 x 24 	 xy

moore.sweetcorn.uniformity 	 24 x 12 	 xy

moore.carrot.uniformity 	 24 x 12 	 xy

moore.springcauliflower.uniformity 	 12 x 20 	 xy

moore.fallcauliflower.uniformity 	 12 x 20 	 xy

nagai.strawberry.uniformity 	 18 x 24 	 xy

nair.turmeric.uniformity 	 72 x 12 	 xy

narain.sorghum.uniformity 	 10 x 16 	 xy

nonnecke.peas.uniformity 	 15 x 18 	 xy, 2 traits

nonnecke.sweetcorn.uniformity 	 32 x 18 	 xy, 3 loc

obsi.potato.uniformity 	 63 x 26, 49 x 19 	 xy

odland.soybean.uniformity 	 25 x 42 	 xy

odland.soyhay.uniformity 	 28 x 55 	 xy

oliveira.lupin.uniformity 	 10 x 40 	 xy

oliveira.potato.uniformity 	 5 x 40 	 xy

paez.coffee.uniformity 	 32 x 36 	 xy, 5 yr

panse.cotton.uniformity 	 32 x 40 	 xy

parker.orange.uniformity 	 10 x 27 	 xy, 6 yr

piepho.barley.uniformity 	 30 x 36 	 xy

polson.safflower.uniformity 	 52 x 33 	 xy 	 smith

rau.coffee.uniformity 	 12 x 12 	 xy

robinson.peanut.uniformity 	 16 x 36 	 xy

robinson.peanut.uniformity 	 16 x 36 	 xy

roemer.sugarbeet.uniformity 	 2 x 48 	 xy

saunders.maize.uniformity 	 5 x 300 	 xy, 2 year

sawyer.multi.uniformity 	 8 x 6 	 xy, 3 year

sayer.sugarcane.uniformity 	 8 x 136, 8 x 121 	 xy, 2 year

shafi.tomato.uniformity 	 10 x 20 	 xy

siao.cotton.uniformity 	 12 x 22 	 xy

smith.beans.uniformity 	 18 x 12, 16 x 15 	 xy, 2 yr, 2 crops

smith.corn.uniformity 	 6 x 20 	 xy, 3 years 	 rgl

smith.wheat.uniformity 	 36 x 30 	 xy, 2 traits

stephens.sorghum.uniformity 	 100 x 20 	 xy

stickler.sorghum.uniformity 	 20 x 20 	 xy, 4 expts, 2 years

strickland.apple.uniformity 	 11 x 18 	 xy

strickland.grape.uniformity 	 5 x 31 	 xy

strickland.peach.uniformity 	 8 x 18 	 xy

strickland.tomato.uniformity 	 30 x 6 	 xy

summerby.multi.uniformity 	 35 x 5 	 xy 	 5 years

tulaikow.wheat.uniformity 	 24 x 10, 15 x 16 	 xy 	 5 years

vishnaadevi.rice.uniformity 	 12 x 24 	 xy

wassom.brome.uniformity 	 36 x 36 	 xy, 3 expts

wehner.cucumber.uniformity 	 6 x 25 	 xy

wiebe.wheat.uniformity 	 12 x 125 	 xy 	 medianpolish, loess

wiedemann.safflower.uniformity 	 54 x 33 	 xy 	 smith

williams.barley.uniformity 	 48 x 15 	 xy 	 loess

williams.cotton.uniformity 	 24 x 12 	 xy 	 loess

wyatt.multi.uniformity 	 43 x 3 	 xy

Yield monitor

name 	 reps	 years 	 trt 	 other 	 model

gartner.corn 	 	 	 	 xy,ym

lasrosas.corn 	 3 	 2 	 6 	 xy,ym 	 lm

kayad.alfalfa 	 	 4 	 	 xy,ym

Animals

name 	 gen 	 years 	 trt 	 other 	 model

alwan.lamb 	 34 	 2 	 	 ordinal 	 clmm

becker.chicken 	 5,12 	 	 	 heritability 	 lmer

crampton.pig 	 	 	 5 	 2 cov	 lm

brandt.switchback 	 10 	 	 2 	 	 aov

depalluel.sheep 	 4 	 	 4 	 latin

diggle.cow 	 	 	 4 	 ts

foulley.calving 	 	 	 	 ordinal 	 polr

goulden.eggs 	 	 	 	 controlchart

harvey.lsmeans 	 3,3 	 	 	 	 lm

harville.lamb 	 5 	 	 	 	 lmer

henderson.milkfat 	 	 	 	 	 nls,lm,glm,gam

holland.arthropods 	 5

ilri.sheep 	 4 	 6 	 	 diallel 	 lmer, asreml

kenward.cattle 	 	 	 2 	 	 asreml

lucas.switchback 	 12 	 	 3 	 	 aov

mead.lamb 	 3 	 	 3 	 	 glm

patterson.switchback 	 12 	 	 4 	 	 aov

urquhart.feedlot 	 11 	 	 3 	 	 lm

woodman.pig 	 	 	 3 	 cov 	 lm

zuidhof.broiler 	 	 	 	 ts

Trees

name 	 gen 	 loc	 reps	 years 	 trt 	 other 	 model

box.cork 	 	 	 	 	 	 repeated 	 radial, asreml

devries.pine 	 	 	 4 	 	 3,3 	 xy,graeco 	 aov

harris.wateruse 	 2 	 	 	 	 2 	 repeated 	 asreml,lme

hanover.whitepine 	 7*4 	 	 4 	 	 	 heritability 	 lmer

johnson.douglasfir 	 	 	 	 	 	 xy

lavoranti.eucalyptus 	 70 	 7 	 	 	 	 	 svd

pearce.apple 	 	 	 4 	 	 6 	 cov 	 lm,lmer

williams.trees 	 37 	 6 	 	 	 2

Field and horticulture crops

name 	 gen 	 loc	 reps	 years 	 trt 	 other 	 model

acorsi.grayleafspot 	 36 	 9 	 2 	 5 	 	 nonnormal 	 gnm,ammi

adugna.sorghum 	 28 	 13 	 	 5

aastveit.barley 	 15 	 	 	 9 	 	 yr*gen~yr*trt	 pls

allcroft.lodging 	 32 	 7 	 	 	 	 percent 	 tobit

archbold.apple 	 2 	 	 5 	 	 24 	 split-split 	 lmer

ars.earlywhitecorn96 	 60 	 9 	 	 	 	 6 traits 	 dotplot

australia.soybean 	 58 	 4 	 	 2 	 	 4-way, 6 traits 	 biplot

bachmaier.nitrogen 	 	 	 4 	 	 2,11 	 	 quadratic lm

barrero.maize 	 847 	 16 	 4 	 11 	 6 	 	 gain,asreml

battese.survey 	 	 12 	 1-5 	 	 2 	 	 lmer

beall.webworms 	 	 	 15 	 	 2,2 	 xy, split-block 	 glm poisson,nb

beaven.barley 	 8 	 	 20 	 	 	 xy

belamkar.augmented 	 273 	 8 	 	 	 	 xy, incblock	 asreml

besag.bayesian 	 75 	 	 3 	 	 	 xy 	 asreml

besag.beans 	 6 	 	 4*6 	 	 	 xy 	 lm,competition

besag.checks 	 2 	 	 	 	 	 xy

besag.elbatan 	 50 	 	 3 	 	 	 xy 	 lm, gam

besag.endive 	 	 	 	 	 	 xy,binary 	 autologistic

besag.met 	 64 	 6 	 3 	 	 	 xy, incblock	 asreml, lme

besag.triticale 	 3 	 	 	 	 2,2,3 	 xy 	 lm, asreml

bliss.borers 	 	 	 	 	 4 	 	 glm

blackman.wheat 	 12 	 7 	 	 	 2 	 	 biplot

bond.diallel 	 6*6 	 	 	 	 9 	 diallel

bridges.cucumber 	 4 	 2 	 4 	 	 	 xy, latin, hetero	 asreml

brandle.rape 	 5 	 9 	 	 3 	 	 	 lmer

buntaran.wheat 	 30 	 18 	 2 	 	 	 alpha 	 asreml

burgueno.alpha 	 15 	 	 3 	 	 	 xy, alpha 	 asreml,lmer

burgueno.rowcol 	 64 	 	 2 	 	 	 xy, row-col 	 asreml,lmer

burgueno.unreplicated 	 280 	 	 	 	 	 xy 	 asreml

butron.maize 	 49 	 3 	 	 2 	 	 diallel,pedigree 	 biplot,asreml

caribbean.maize 	 	 17 	 4 	 	 3

carmer.density 	 8 	 	 	 	 4 	 	 nls,nlme

carlson.germination 	 15 	 	 	 	 8 	 	 glm

chakravertti.factorial 	 3 	 	 3 	 	 3,5,3,3 	 factorial 	 aov

chinloy.fractionalfactorial 	 	 	 9 	 	 1/3 3^5 = 3,3,3,3 	 xy,factorial 	 aov

christidis.competition 	 9 	 	 5 	 	 	 xy

cochran.beets 	 	 	 6 	 	 7

cochran.bib 	 13 	 13 	 	 	 	 bib 	 aov, lme

cochran.crd 	 	 	 	 	 7 	 xy, crd	 aov

cochran.factorial 	 	 	 2 	 	 2,2,2,2 = 2^4 	 factorial 	 aov

cochran.latin 	 	 	 6 	 	 6 	 xy, latin	 aov

cochran.lattice 	 	 	 5 	 	 16 	 xy, latin	 lmer

cochran.wireworms 	 	 	 5 	 	 5 	 xy, latin	 glm

cochran.eelworms 	 	 	 4 	 	 5 	 xy 	 aov

connolly.potato 	 20 	 	 4 	 	 	 xy, competition 	 lm

cornelius.maize 	 9 	 20 	 	 	 	 	 svd

corsten.interaction 	 20 	 7

cramer.cucumber 	 	 	 8 	 	 	 	 pathcoef

crossa.wheat 	 18 	 25 	 	 	 	 	 ammi

crowder.seeds 	 2 	 	 21 	 	 2 	 	 glm,INLA,jags

cox.stripsplit 	 	 	 4 	 	 3,4,2 	 split-block 	 aov

cullis.earlygen 	 532 	 	 	 	 	 xy 	 asreml

damesa.maize 	 22 	 4 	 3 	 	 	 xy,incblock,twostage 	 asreml

dasilva.maize 	 55 	 9 	 3

darwin.maize 	 	 	 12 	 	 2 	 	 t.test

davidian.soybean 	 2 	 	 	 3 	 	 	 nlme

denis.missing 	 5 	 26 	 	 	 	 	 lme

denis.ryegrass 	 21 	 7 	 	 	 	 	 aov

digby.jointregression 	 10 	 17 	 	 	 4 	 	 lm

durban.competition 	 36 	 	 3 	 	 	 xy, competition 	 lm

durban.rowcol 	 272 	 	 2 	 	 	 xy 	 lm, gam, asreml

durban.splitplot 	 70 	 	 4 	 	 2 	 xy 	 lm, gam, asreml

eden.potato 	 	 	 4 	 3 	 4-12 	 xy, rcb, latin 	 aov

eden.nonnormal 	 	 	 4 	 	 4 	 	 aov

edwards.oats 	 80 	 5 	 3 	 7

engelstad.nitro 	 	 2 	 	 5 	 6 	 rsm1 	 nls quadratic plateau

fan.stability 	 13 	 10 	 	 2 	 	 3-way 	 stability

federer.diagcheck 	 122 	 	 	 	 	 xy 	 lm, lmer, asreml

federer.tobacco 	 	 	 8 	 	 7 	 xy 	 lm

fisher.barley 	 5 	 6 	 	 2

fisher.latin 	 	 	 5 	 	 5 	 xy,latin 	 lm

fox.wheat 	 22 	 14 	 	 	 	 	 lm

gathmann.bt 	 2 	 	 8 	 	 	 	 tost

gauch.soy 	 7 	 7 	 4 	 12 	 	 	 ammi

george.wheat 	 211 	 9 	 4 	 15

giles.wheat 	 19 	 	 	 13 	 	 2 traits	 gnm

gilmour.serpentine 	 108 	 	 3 	 	 	 xy, serpentine	 asreml

gilmour.slatehall 	 25 	 	 6 	 	 	 xy 	 asreml

gomez.fractionalfactorial 	 	 	 2 	 	 1/2 2^6 = 2,2,2,2,2,2 	 xy,factorial 	 lm

gomez.groupsplit 	 45 	 	 3 	 	 2 	 xy, 3 gen groups 	 aov

gomez.heteroskedastic 	 35 	 	 3 	 	 	 hetero

gomez.multilocsplitplot 	 2 	 3 	 3 	 	 	 rsm1,nitro 	 aov, lmer

gomez.nitrogen 	 	 	 4 	 	 8 	 	 aov, contrasts

gomez.nonnormal1 	 	 	 4 	 	 9 	 log10 	 lm

gomez.nonnormal2 	 14 	 	 3 	 	 	 sqrt 	 lm

gomez.nonnormal3 	 12 	 	 3 	 	 	 arcsin 	 lm

gomez.seedrate 	 	 	 4 	 	 6 	 rate 	 lm

gomez.splitplot.subsample 	 	 	 3 	 	 8,4 	 subsample 	 aov

gomez.splitsplit 	 3 	 	 3 	 	 	 xy, nitro, mgmt 	 aov, lmer

gomez.stripplot 	 6 	 	 3 	 	 	 xy, nitro 	 aov

gomez.stripsplitplot 	 6 	 	 3 	 	 	 xy, nitro	 aov

gomez.wetdry 	 	 	 3 	 2 	 5 	 nitro	 lmer

gotway.hessianfly 	 16 	 	 4 	 	 	 xy 	 lmer

goulden.latin 	 	 	 5 	 	 5 	 xy, latin 	 lm

goulden.splitsplit 	 2 	 	 4 	 	 2*5 	 xy, split 	 aov

graybill.heteroskedastic 	 4 	 13 	 	 	 	 hetero

gregory.cotton 	 	 	 	 2 	 4*3*2*2 	 polar

grover.diallel 	 	 	 4 	 	 6*6 	 diallel 	 lmDiallel

grover.rcb.subsample 	 	 	 4 	 2 	 9 	 subsample 	 aov

gumpertz.pepper 	 	 	 	 	 	 xy 	 glm

hadasch.lettuce 	 89 	 3 	 3 	 	 	 markers 	 asreml

hanks.sprinkler 	 3 	 	 3 	 	 	 xy 	 asreml

hayman.tobacco 	 8 	 	 2 	 2 	 	 diallel	 asreml

hazell.vegetables 	 4 	 	 	 6 	 	 	 linprog

heady.fertilizer 	 	 	 2 	 	 9*9 	 rsm2 	 lm,rgl

hernandez.nitrogen 	 	 5 	 4 	 	 	 rsm1 	 lm, nls

hildebrand.systems 	 	 14 	 	 	 4 	 	 asreml

holshouser.splitstrip 	 4 	 	 4 	 	 2*4 	 rsm1,pop 	 lmer

huehn.wheat 	 20 	 10 	 	 	 	 	 huehn

hughes.grapes 	 	 	 3 	 	 6 	 binomial 	 lmer, aod, glmm

hunter.corn 	 	 12 	 	 3 	 1 	 rsm1 	 xyplot

ivins.herbs 	 13 	 	 6 	 	 	 2 traits 	 lm, friedman

jansen.apple 	 3 	 	 4 	 	 3 	 binomial 	 glmer

jansen.carrot 	 16 	 	 3 	 	 2 	 binomial 	 glmer

jansen.strawberry 	 12 	 	 4 	 	 	 ordinal 	 mosaicplot

jayaraman.bamboo 	 6 	 2 	 3 	 	 	 heritability 	 lmer

jenkyn.mildew 	 	 	 9 	 	 4 	 	 lm

john.alpha 	 24 	 	 3 	 	 	 xy, alpha 	 lm, lmer

johnson.blight 	 	 	 	 2 	 	 	 logistic

kang.maize 	 17 	 4 	 	 3 	 2,4

kang.peanut 	 10 	 15 	 4 	 	 	 	 gge

karcher.turfgrass 	 	 	 4 	 	 2,4 	 ordinal 	 polr

keen.potatodamage 	 6 	 	 4 	 	 2,3,8 	 ordinal 	 mosaicplot,clmm

kempton.competition 	 36 	 	 3 	 	 	 xy, competition 	 lme AR1

kempton.rowcol 	 35 	 	 2 	 	 	 xy, row-col 	 lmer

kling.augmented 	 53 	 	 6 	 	 	 xy, augmented 	 lmer

kempton.slatehall 	 25 	 	 6 	 	 	 xy 	 asreml, lmer

kirk.potato 	 21 	 	 15 	 	 	 xy

lee.potatoblight 	 337 	 	 4 	 11 	 	 xy, ordinal, repeated

lehner.soybeanmold 	 	 35 	 	 4 	 11 	 	 metafor, lmer

lillemo.wheat 	 24 	 13 	 	 7 	 	 qq 	 medpolish, huehn

lin.superiority 	 33 	 12 	 	 	 	 	 superiority

lin.unbalanced 	 33 	 18 	 	 	 	 	 superiority

linder.wheat 	 9 	 7 	 4 	 	 	 	 gge

little.splitblock 	 	 	 4 	 	 4,5 	 xy, split-block	 aov

lonnquist.maize 	 11 	 	 	 	 	 diallel 	 asreml

lyons.wheat 	 	 12 	 	 4

lu.stability 	 5 	 6 	 	 	 	 	 huehn

mcconway.turnip 	 2 	 	 4 	 	 2,4 	 hetero 	 aov, lme

mcleod.barley 	 	 8 	 	 	 6 	 	 aggregate

mead.cauliflower 	 	 	 	 2 	 	 poisson 	 glm

mead.cowpea.maize 	 3,2 	 	 3 	 	 4 	 intercrop

mead.germination 	 	 	 4 	 	 4,4 	 binomial	 glm

mead.strawberry 	 8 	 	 4

mead.turnip 	 	 	 3 	 	 5,4 	 	 aov

miguez.biomass 	 3 	 	 4

minnesota.barley.weather 	 	 6 	 	 10

minnesota.barley.yield 	 22 	 6 	 	 10 	 	 	 dotplot

omer.sorghum 	 18 	 6 	 4 	 	 	 	 jags

onofri.winterwheat 	 8 	 	 3 	 7 	 	 	 ammi

ortiz.tomato 	 15 	 18 	 	 	 16 	 env*gen~env*cov 	 pls

pacheco.soybean 	 18 	 11 	 	 	 	 	 ammi

payne.wheat 	 	 	 	 20 	 6 	 rotation	 asreml

pederson.lettuce.repeated 	 	 	 18 	 	 3 	 	 nlme

perry.springwheat 	 28 	 5 	 	 4 	 	 gain 	 lm,lmer,asreml

petersen.sorghum.cowpea 	 2 	 	 4 	 7 	 4 	 intercrop

piepho.cocksfoot 	 25 	 	 	 7 	 	 	 mumm

ratkowsky.onions 	 	 	 	 	 	 	 lm

reid.grasses 	 4 	 	 	 3 	 21 	 	 nlme SSfpl

riddle.wheat 	 25 	 	 5 	 2 	 	 xy, latin 	 aov

ridout.appleshoots 	 	 	 30 	 	 2,4 	 zip 	 zeroinfl

rothamsted.brussels 	 	 	 4 	 	 6

rothamsted.oats 	 	 	 8 	 	 9 	 rcb

ryder.groundnut 	 5 	 	 4 	 	 	 xy, rcb 	 lm

salmon.bunt 	 10 	 	 2 	 	 20 	 	 betareg

senshu.rice 	 	 40 	 	 	 	 	 lm,Fieller

shafii.rapeseed 	 6 	 14 	 3 	 3 	 	 	 biplot

shaw.oats 	 13 	 2 	 5 	 3 	 	 	 aov

sharma.met 	 7 	 3 	 3 	 2 	 	 	 FinlayWilkinson

silva.cotton 	 	 	 5 	 	 5 	 5 traits 	 glm,poisson

sinclair.clover 	 	 	 	 	 5,5 	 rsm2,mitzerlich 	 nls,rgl

snedecor.asparagus 	 	 	 4 	 4 	 4 	 	 split-plot, antedependence

snijders.fusarium 	 17 	 	 	 3 	 4 	 percent 	 glm/gnm,gammi

steptoe.morex.pheno 	 152 	 16 	 	 	 	 10 traits

steptoe.morex.geno 	 150 	 	 	 	 	 223 markers, qtl

streibig.competition 	 2 	 	 3 	 	 	 	 glm

stroup.nin 	 56 	 	 4 	 	 	 xy 	 asreml

stroup.splitplot 	 	 	 4 	 	 	 	 asreml, MCMCglmm

student.barley 	 2 	 51 	 	 6 	 	 	 lmer

tai.potato 	 8 	 3 	 	 2 	 	 	 tai

talbot.potato 	 9 	 12 	 	 	 	 gen*env~gen*trt 	 pls

tesfaye.millet 	 47 	 2 	 2-3 	 2 	 4 	 xy 	 asreml

theobald.barley 	 3 	 5 	 	 2 	 5 	 rsm1

theobald.covariate 	 10 	 7 	 5 	 	 	 cov 	 jags

thompson.cornsoy 	 	 5 	 	 33 	 	 repeated measures 	 aov

vaneeuwijk.fusarium 	 20 	 	 	 4 	 7 	 3-way 	 aov

vaneeuwijk.drymatter 	 6 	 4 	 	 7 	 	 3-way 	 aov,lmer

vaneeuwijk.nematodes 	 11 	 	 	 	 	 nonnormal,poisson 	 gnm, gammi

vargas.wheat1 	 7 	 	 	 6 	 	 gen*yr~gen*trt, yr*gen~yr*cov	 pls

vargas.wheat2 	 8 	 7 	 	 	 	 env*gen~env*cov 	 pls

vargas.txe 	 	 	 	 10 	 24 	 yr*trt~yr*cov 	 pls

verbyla.lupin 	 9 	 8 	 3 	 2 	 7 	 rsm1, xy, density	 asreml

vold.longterm 	 	 	 	 19 	 4 	 rsm1 	 nls,nlme

vsn.lupin3 	 336 	 3 	 	 	 	 xy 	 asreml

wedderburn.barley 	 10 	 9 	 	 	 	 percent 	 glm/gnm

weiss.incblock 	 31 	 	 6 	 	 	 xy,incblock 	 asreml

weiss.lattice 	 49 	 	 4 	 	 	 xy,lattice 	 lm,asreml

welch.bermudagrass 	 	 	 	 	 4,4,4 	 rsm3, factorial 	 lm, jags

wheatley.carrot 	 	 	 3 	 	 11 	 	 glm-binomial

yan.winterwheat 	 18 	 9 	 	 	 	 	 gge,biplot

yang.barley 	 6 	 18 	 	 	 	 	 biplot

yates.missing 	 	 	 10 	 	 3^2 = 3,3 	 factorial 	 lm, pca

yates.oats 	 3 	 	 6 	 	 	 xy,split,nitro	 lmer

Time series

name 	 years 	 trt 	 other 	 model

byers.apple 	 	 	 	 lme

broadbalk.wheat 	 74 	 17

hessling.argentina 	 30 	 	 temp,precip

kreusler.maize 	 4 	 5 	 plant growth

lambert.soiltemp 	 1 	 7

nass.barley 	 146

nass.corn 	 146

nass.cotton 	 146

nass.hay 	 104

nass.sorghum 	 93

nass.wheat 	 146

nass.rice 	 117

nass.soybean 	 88

usda.gmoadoption 	 31 	 5

walsh.cottonprice 	 34 	 	 	 cor

Other

name 	 model

cate.potassium 	 cate-nelson

cleveland.soil 	 loess 2D

harrison.priors 	 nls, prior

nebraska.farmincome 	 choropleth

pearl.kernels 	 chisq

stirret.borers 	 lm, 4 trt

turner.herbicide 	 glm, 4 trt

usgs.herbicides 	 non-detect

wallace.iowaland 	 lm, choropleth

waynick.soil 	 spatial, nitro/carbon

Summaries:

Diallel experiments:

name 	 gen 	 loc	 reps	 trt 	 model

bond.diallel 	 6*6 	 	 	 9

butron.maize 	 49 	 3 	 	 	 biplot,asreml

grover.diallel 	 	 	 4 	 6*6 	 lmDiallel

hayman.tobacco 	 8 	 	 2 	 	 asreml

ilri.sheep 	 4 	 6

lonnquist.maize 	 11 	 	 	 	 asreml

Factorial experiments:

name 	 gen 	 loc	 reps	 years 	 trt 	 other 	 model

chakravertti.factorial 	 3 	 	 3 	 	 3,5,3,3 	 factorial 	 aov

chinloy.fractionalfactorial 	 	 	 9 	 	 1/3 3^5 = 3,3,3,3 	 xy,factorial 	 aov

cochran.factorial 	 	 	 2 	 	 2,2,2,2 = 2^4 	 factorial 	 aov

gomez.fractionalfactorial 	 	 	 2 	 	 1/2 2^6 = 2,2,2,2,2,2 	 xy,factorial 	 lm

welch.bermudagrass 	 	 	 	 	 4,4,4 	 rsm3, factorial 	 lm, jags

yates.missing 	 	 	 10 	 	 3^2 = 3,3 	 factorial 	 lm, pca

Multi-environment trials with multi-genotype,loc,rep,year:

name 	 gen 	 loc	 reps	 years 	 trt 	 other 	 model

barrero.maize 	 847 	 16 	 4 	 11 	 6 	 	 asreml

edwards.oats 	 80 	 5 	 3 	 7

gauch.soy 	 7 	 7 	 4 	 12 	 	 	 ammi

george.wheat 	 211 	 9 	 4 	 15

shafii.rapeseed 	 6 	 14 	 3 	 3 	 	 	 biplot

shaw.oats 	 13 	 2 	 5 	 3 	 	 	 aov

tesfaye.millet 	 47 	 2 	 2-3 	 2 	 4 	 xy,FA 	 asreml

verbyla.lupin 	 9 	 8 	 3 	 2 	 7 	 rsm1, xy, density 	 asreml

Data with markers: hadasch.lettuce.markers, steptoe.morex.geno

Data with pedigree: butron.maize

Author(s)

Kevin Wright, with support from many people who granted permission to include their data in this package.

References

J. White and Frits van Evert. (2008). Publishing Agronomic Data. Agron J. 100, 1396-1400. https://doi.org/10.2134/agronj2008.0080F

[Package agridat version 1.26 Index]