Diallel 6x6

Diallel 6x6 in 4 blocks.

data("grover.diallel")

Format

A data frame with 144 observations on the following 5 variables.

yield

a numeric vector

rep

a character vector

parent1

a character vector

parent2

a character vector

cross

a character vector

Details

Yield for a 6x6 diallel with 4 reps.

Note: The mean for the 2x2 cross is slightly different than Grover p. 252. There appears to be an unknown error in the one of the 4 reps in the data on page 250.

Source

Grover, Deepak & Lajpat Rai (2010). Experimental Designing And Data Analysis In Agriculture And Biology. Agrotech Publishing Academy. Page 85. https://archive.org/details/expldesnanddatanalinagblg00023

References

None

Examples

# \dontrun{
  data(grover.diallel)
  dat <- grover.diallel

  anova(aov(yield ~ rep + cross, data=dat))
#> Analysis of Variance Table
#> 
#> Response: yield
#>            Df Sum Sq Mean Sq F value    Pr(>F)    
#> rep         3   2034  677.86  5.7478  0.001113 ** 
#> cross      35  32773  936.38  7.9399 < 2.2e-16 ***
#> Residuals 105  12383  117.93                      
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

  # These effects match the GCA and SCA values in Grover table 3, page 253.
  
  library(lmDiallel)
  m2 <- lm.diallel(yield ~ parent1 + parent2, Block=rep,
                   data=dat, fct="GRIFFING1")
  library(multcomp)
#> Loading required package: mvtnorm
#> Loading required package: survival
#> 
#> Attaching package: 'survival'
#> The following object is masked from 'package:asreml':
#> 
#>     rats
#> Loading required package: TH.data
#> 
#> Attaching package: 'TH.data'
#> The following object is masked from 'package:MASS':
#> 
#>     geyser
  summary( glht(linfct=diallel.eff(m2), test=adjusted(type="none")) )
#> Warning: Completion with error > abseps
#> Warning: Completion with error > abseps
#> Warning: Completion with error > abseps
#> Warning: Completion with error > abseps
#> Warning: Completion with error > abseps
#> Warning: Completion with error > abseps
#> Warning: Completion with error > abseps
#> Warning: Completion with error > abseps
#> Warning: Completion with error > abseps
#> Warning: Completion with error > abseps
#> Warning: Completion with error > abseps
#> Warning: Completion with error > abseps
#> 
#> 	 Simultaneous Tests for General Linear Hypotheses
#> 
#> Linear Hypotheses:
#>                Estimate Std. Error t value Pr(>|t|)    
#> Intercept == 0  93.0774     0.9050 102.851    <0.01 ***
#> g_P1 == 0        1.4851     1.4309   1.038   1.0000    
#> g_P2 == 0       -0.9911     1.4309  -0.693   1.0000    
#> g_P3 == 0        2.2631     1.4309   1.582   0.9749    
#> g_P4 == 0        5.4247     1.4309   3.791   0.0303 *  
#> g_P5 == 0       -4.2490     1.4309  -2.969   0.1970    
#> g_P6 == 0       -3.9328     1.4309  -2.748   0.3000    
#> ts_P1:P1 == 0  -10.4026     4.5249  -2.299   0.6025    
#> ts_P1:P2 == 0   -9.7214     3.2629  -2.979   0.1931    
#> ts_P1:P3 == 0   -0.4581     3.2629  -0.140   1.0000    
#> ts_P1:P4 == 0   17.0428     3.2629   5.223    <0.01 ***
#> ts_P1:P5 == 0   25.4765     3.2629   7.808    <0.01 ***
#> ts_P1:P6 == 0  -21.9372     3.2629  -6.723    <0.01 ***
#> ts_P2:P1 == 0   -9.7214     3.2629  -2.979   0.1925    
#> ts_P2:P2 == 0    7.0899     4.5249   1.567   0.9772    
#> ts_P2:P3 == 0   13.7157     3.2629   4.203   0.0110 *  
#> ts_P2:P4 == 0   -8.7710     3.2629  -2.688   0.3351    
#> ts_P2:P5 == 0   13.5728     3.2629   4.160   0.0124 *  
#> ts_P2:P6 == 0  -15.8860     3.2629  -4.869    <0.01 ** 
#> ts_P3:P1 == 0   -0.4581     3.2629  -0.140   1.0000    
#> ts_P3:P2 == 0   13.7157     3.2629   4.203   0.0108 *  
#> ts_P3:P3 == 0  -23.5335     4.5249  -5.201    <0.01 ***
#> ts_P3:P4 == 0    1.2699     3.2629   0.389   1.0000    
#> ts_P3:P5 == 0    0.1836     3.2629   0.056   1.0000    
#> ts_P3:P6 == 0    8.8224     3.2629   2.704   0.3256    
#> ts_P4:P1 == 0   17.0428     3.2629   5.223    <0.01 ***
#> ts_P4:P2 == 0   -8.7710     3.2629  -2.688   0.3352    
#> ts_P4:P3 == 0    1.2699     3.2629   0.389   1.0000    
#> ts_P4:P4 == 0  -12.2868     4.5249  -2.715   0.3193    
#> ts_P4:P5 == 0   -9.9156     3.2629  -3.039   0.1707    
#> ts_P4:P6 == 0   12.6607     3.2629   3.880   0.0245 *  
#> ts_P5:P1 == 0   25.4765     3.2629   7.808    <0.01 ***
#> ts_P5:P2 == 0   13.5728     3.2629   4.160   0.0123 *  
#> ts_P5:P3 == 0    0.1836     3.2629   0.056   1.0000    
#> ts_P5:P4 == 0   -9.9156     3.2629  -3.039   0.1705    
#> ts_P5:P5 == 0  -30.4793     4.5249  -6.736    <0.01 ***
#> ts_P5:P6 == 0    1.1619     3.2629   0.356   1.0000    
#> ts_P6:P1 == 0  -21.9372     3.2629  -6.723    <0.01 ***
#> ts_P6:P2 == 0  -15.8860     3.2629  -4.869    <0.01 ** 
#> ts_P6:P3 == 0    8.8224     3.2629   2.704   0.3259    
#> ts_P6:P4 == 0   12.6607     3.2629   3.880   0.0244 *  
#> ts_P6:P5 == 0    1.1619     3.2629   0.356   1.0000    
#> ts_P6:P6 == 0   15.1782     4.5249   3.354   0.0851 .  
#> r_P1:P2 == 0     3.1600     3.8395   0.823   1.0000    
#> r_P1:P3 == 0    -5.8625     3.8395  -1.527   0.9831    
#> r_P1:P4 == 0    -2.0850     3.8395  -0.543   1.0000    
#> r_P1:P5 == 0     4.5000     3.8395   1.172   0.9996    
#> r_P1:P6 == 0    -0.1425     3.8395  -0.037   1.0000    
#> r_P2:P1 == 0    -3.1600     3.8395  -0.823   1.0000    
#> r_P2:P3 == 0     3.5100     3.8395   0.914   1.0000    
#> r_P2:P4 == 0    -0.5700     3.8395  -0.148   1.0000    
#> r_P2:P5 == 0    -1.4800     3.8395  -0.385   1.0000    
#> r_P2:P6 == 0     0.9725     3.8395   0.253   1.0000    
#> r_P3:P1 == 0     5.8625     3.8395   1.527   0.9832    
#> r_P3:P2 == 0    -3.5100     3.8395  -0.914   1.0000    
#> r_P3:P4 == 0    -0.6400     3.8395  -0.167   1.0000    
#> r_P3:P5 == 0     3.0100     3.8395   0.784   1.0000    
#> r_P3:P6 == 0     0.6550     3.8395   0.171   1.0000    
#> r_P4:P1 == 0     2.0850     3.8395   0.543   1.0000    
#> r_P4:P2 == 0     0.5700     3.8395   0.148   1.0000    
#> r_P4:P3 == 0     0.6400     3.8395   0.167   1.0000    
#> r_P4:P5 == 0     0.9475     3.8395   0.247   1.0000    
#> r_P4:P6 == 0    -1.4350     3.8395  -0.374   1.0000    
#> r_P5:P1 == 0    -4.5000     3.8395  -1.172   0.9996    
#> r_P5:P2 == 0     1.4800     3.8395   0.385   1.0000    
#> r_P5:P3 == 0    -3.0100     3.8395  -0.784   1.0000    
#> r_P5:P4 == 0    -0.9475     3.8395  -0.247   1.0000    
#> r_P5:P6 == 0    -1.9075     3.8395  -0.497   1.0000    
#> r_P6:P1 == 0     0.1425     3.8395   0.037   1.0000    
#> r_P6:P2 == 0    -0.9725     3.8395  -0.253   1.0000    
#> r_P6:P3 == 0    -0.6550     3.8395  -0.171   1.0000    
#> r_P6:P4 == 0     1.4350     3.8395   0.374   1.0000    
#> r_P6:P5 == 0     1.9075     3.8395   0.497   1.0000    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> (Adjusted p values reported -- single-step method)
#> 
  ## Linear Hypotheses:
  ##                Estimate Std. Error t value Pr(>|t|)    
  ## Intercept == 0  93.0774     0.9050 102.851    <0.01 ***
  ## g_P1 == 0        1.4851     1.4309   1.038   1.0000    
  ## g_P2 == 0       -0.9911     1.4309  -0.693   1.0000    
  ## g_P3 == 0        2.2631     1.4309   1.582   0.9748    
  ## g_P4 == 0        5.4247     1.4309   3.791   0.0302 *  
  ## g_P5 == 0       -4.2490     1.4309  -2.969   0.1972    
  ## g_P6 == 0       -3.9328     1.4309  -2.748   0.3008    
  ## ts_P1:P1 == 0  -10.4026     4.5249  -2.299   0.6014    
  ## ts_P1:P2 == 0   -9.7214     3.2629  -2.979   0.1933    
  ## ts_P1:P3 == 0   -0.4581     3.2629  -0.140   1.0000    
  ## ts_P1:P4 == 0   17.0428     3.2629   5.223    <0.01 ***
  ## ts_P1:P5 == 0   25.4765     3.2629   7.808    <0.01 ***
  ## ts_P1:P6 == 0  -21.9372     3.2629  -6.723    <0.01 ***
  ## ts_P2:P1 == 0   -9.7214     3.2629  -2.979   0.1928    
  ## ts_P2:P2 == 0    7.0899     4.5249   1.567   0.9773    

# }