libs {agridat}	R Documentation

Load multiple packages and install if needed

Description

Install and load packages "on the fly".

Usage

libs(...)

Arguments

...
Comma-separated unquoted package names

Details

The 'agridat' package uses dozens of packages in the examples for each dataset. The 'libs' function provides a simple way to load multiple packages at once, and can install any missing packages on-the-fly.

This is very similar to the 'pacman::p_load' function.

Value

None

Author(s)

Kevin Wright

References

None

Examples

## Not run: 
libs(dplyr,reshape2)

## End(Not run)

[Package agridat version 1.26 Index]