Uniformity trials of wheat, swedes, oats, 3 years on the same land

Uniformity trials of wheat, swedes, oats at Rothamsted, England, 1925-1927.

Usage

data("sawyer.multi.uniformity")

Format

A data frame with 48 observations on the following 7 variables.

year

year

crop

crop

row

row

col

column

yield

crop yield

Details

An experiment conducted at Rothamsted, England, in Sawyers Field.

Eden & Maskell (page 165) say the field was clover, and ploughed in the autumn of 1924. The field was laid out uniformly in lands of one chain width and each plot width made to coincide with the land width from ridge to ridge. The length of each plot was also one chain and from the point of view of yield data the trial comprised 47 plots in 8x6 except that a hedge only allowed a rank of 6 plots at one end. Row 6, column 1 was not planted in any year.

Field width: 8 plots * 22 feet = 528 feet

Field length: 6 plots * 22 feet = 396 feet

The field is 8 plots wide, 6 plots long. The plots are drawn in the source documents as squares .098 acres each (1 chain = 66 feet on each side).

1925: Wheat was harvested. Grain and straw weight measured in pounds.

Row 1, column 1 had partially missing data for the wheat values in 1925 and was not used in the Rothamsted summary statistics on page 155.

1926: Swedes were harvested. Leaf weight and root weight measured in pounds. Root count was measured. Note: Swedes are also called rutabagas.

1927: Oats were harvested. Grain and straw weight measured.

Note the summaries statistics at the bottom of the page in each report are calibrated to ACRES.

Source

Rothamsted Experimental Station, Report 1925-26. Lawes Agricultural Trust, p. 154-155. https://www.era.rothamsted.ac.uk/eradoc/book/84

Rothamsted Experimental Station, Report 1927-1928. Lawes Agricultural Trust, p. 153. https://www.era.rothamsted.ac.uk/eradoc/article/ResReport1927-28-131-175

References

Eden, T. and E. J. Maskell. (1928). The influence of soil heterogeneity on the growth and yield of successive crops. Jour of Agricultural Science, 18, 163-185. https://archive.org/stream/in.ernet.dli.2015.25895/2015.25895.Journal-Of-Agricultural-Science-Vol-xviii-1928#page/n175

McCullagh, P. and Clifford, D., (2006). Evidence for conformal invariance of crop yields, Proceedings of the Royal Society A: Mathematical, Physical and Engineering Science, 462, 2119–2143. https://doi.org/10.1098/rspa.2006.1667

Winifred A. Mackenzie. (1926) Note on a remarkable correlation between grain and straw, obtained at Rothamsted. Journal of Agricultural Science, 16, 275-279. https://doi.org/10.1017/S0021859600018256

Examples

if (FALSE) { # \dontrun{
  
  library(agridat)
  data("sawyer.multi.uniformity")
  dat <- sawyer.multi.uniformity
  
  libs(desplot)
  # The field plan shows square plots
  desplot(dat, yield~col*row,
          subset= (year==1925 & crop=="wheat_grain"),
          main="sawyer.multi.uniformity - 1925 wheat grain yield",
          aspect=(6)/(8)) # true aspect
  
  desplot(dat, yield~col*row,
          subset= (year==1926 & crop=="swedes_rootwt"),
          main="sawyer.multi.uniformity - 1926 root weight of swedes",
          aspect=(6)/(8))

  desplot(dat, yield~col*row,
          subset= (year==1927 & crop=="oats_grain"),
          main="sawyer.multi.uniformity - 1927 oats grain yield",
          aspect=(6)/(8))

  # Match Mackenzie, p. 279. This plot shows a possible "outlier".
  d1 <- subset(dat, year==1925 & crop=="wheat_grain")
  d2 <- subset(dat, year==1925 & crop=="wheat_straw")
  d12 <- merge(d1[ , c("row","col","yield")],
               d2[ , c("row","col","yield")], by=c("row","col"))
  plot(d12$yield.y, d12$yield.x, xlab="1925 straw", ylab="1925 grain")
  
  # Swede correlations, matches McCullagh p 2121
  ##        leafwt rootwt rootct
  ## leafwt   1.00   0.66   0.47
  ## rootwt   0.66   1.00   0.43
  ## rootct   0.47   0.43   1.00
  
} # }